does it work

+++ build_pack.sh
@@ -5,6 +5,11 @@ cd "$(dirname "$0")"
 echo "Compiling..."
 msbuild ./Sonic4_Tools/Sonic4_Tools.sln /p:Configuration=Release
 
+EXIT_CODE="$?"
+if [ "$EXIT_CODE" != "0" ]; then
+    exit $EXIT_CODE
+fi
+
 echo "Removing old distribution package..."
 rm -r "./dist"
 mkdir -p "./dist/Sonic4_Tools/Common/Licenses"
